Singularize folders names
