--- src/Configs/Envs.php
@@ -1,24 +1,0 @@
-<?php
-
-namespace Yiisoft\Composer\Config\Configs;
-
-/**
- * DotEnv class represents output configuration file with ENV values.
- */
-class Envs extends Config
-{
-    public function envsRequired(): bool
-    {
-        return false;
-    }
-
-    public function constantsRequired(): bool
-    {
-        return false;
-    }
-
-    public function paramsRequired(): bool
-    {
-        return false;
-    }
-}
