+++ src/Exception/BadConfigurationException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Yiisoft\Composer\Config\Exception;
+
+/**
+ * Bad configuration exception.
+ */
+class BadConfigurationException extends Exception
+{
+}
