+++ src/Exception/CircularDependencyException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Yiisoft\Composer\Config\Exception;
+
+/**
+ * Circular dependency exception.
+ */
+class CircularDependencyException extends Exception
+{
+}
