+++ src/Exception/Exception.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Yiisoft\Composer\Config\Exception;
+
+/**
+ * Common ancestor for all exception thrown by this library.
+ */
+class Exception extends \Exception
+{
+}
