+++ src/Exception/FailedReadException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Yiisoft\Composer\Config\Exception;
+
+/**
+ * Failed read file exception.
+ */
+class FailedReadException extends Exception
+{
+}
