+++ src/Exception/FailedWriteException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Yiisoft\Composer\Config\Exception;
+
+/**
+ * Failed write file exception.
+ */
+class FailedWriteException extends Exception
+{
+}
