+++ src/Exception/UnsupportedFileTypeException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Yiisoft\Composer\Config\Exception;
+
+/**
+ * Unsupported file type exception.
+ */
+class UnsupportedFileTypeException extends FailedReadException
+{
+}
