--- src/Exceptions/BadConfigurationException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Yiisoft\Composer\Config\Exceptions;
-
-/**
- * Bad configuration exception.
- */
-class BadConfigurationException extends Exception
-{
-}
