--- src/Exceptions/CircularDependencyException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Yiisoft\Composer\Config\Exceptions;
-
-/**
- * Circular dependency exception.
- */
-class CircularDependencyException extends Exception
-{
-}
