--- src/Exceptions/Exception.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Yiisoft\Composer\Config\Exceptions;
-
-/**
- * Common ancestor for all exception thrown by this library.
- */
-class Exception extends \Exception
-{
-}
