--- src/Exceptions/FailedReadException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Yiisoft\Composer\Config\Exceptions;
-
-/**
- * Failed read file exception.
- */
-class FailedReadException extends Exception
-{
-}
