--- src/Exceptions/FailedWriteException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Yiisoft\Composer\Config\Exceptions;
-
-/**
- * Failed write file exception.
- */
-class FailedWriteException extends Exception
-{
-}
