--- src/Exceptions/UnsupportedFileTypeException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Yiisoft\Composer\Config\Exceptions;
-
-/**
- * Unsupported file type exception.
- */
-class UnsupportedFileTypeException extends FailedReadException
-{
-}
