--- src/Readers/ReaderInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Composer\Config\Readers;
-
-interface ReaderInterface
-{
-    public function read(string $path): array;
-}
