+++ src/Util/RemoveArrayKeys.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Yiisoft\Composer\Config\Util;
+
+/**
+ * It's just a marker. Does nothing.
+ */
+class RemoveArrayKeys
+{
+}
