--- src/Utils/RemoveArrayKeys.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Yiisoft\Composer\Config\Utils;
-
-/**
- * It's just a marker. Does nothing.
- */
-class RemoveArrayKeys
-{
-}
